requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
anthropic>=0.40.0
notion-client>=2.0.0
Pillow>=10.0.0
pydub>=0.25.1
httpx>=0.27.0
APScheduler>=3.10.0
PyYAML>=6.0
pytest>=8.0.0
